requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile ../requirements.in
#
asgiref==3.8.1
    # via django
asttokens==2.4.1
    # via stack-data
attrs==24.1.0
    # via
    #   jsonschema
    #   referencing
coverage[toml]==7.6.0
    # via
    #   -r ../requirements.in
    #   pytest-cov
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
django==5.0.7
    # via
    #   -r ../requirements.in
    #   django-extensions
    #   django-js-asset
    #   djangorestframework
    #   drf-spectacular
django-extensions==3.2.3
    # via -r ../requirements.in
django-js-asset==2.2.0
    # via django-mptt
django-mptt==0.16.0
    # via -r ../requirements.in
djangorestframework==3.15.2
    # via
    #   -r ../requirements.in
    #   drf-spectacular
drf-spectacular==0.27.2
    # via -r ../requirements.in
executing==2.0.1
    # via stack-data
factory-boy==3.3.0
    # via pytest-factoryboy
faker==26.1.0
    # via factory-boy
inflection==0.5.1
    # via
    #   drf-spectacular
    #   pytest-factoryboy
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
    # via -r ../requirements.in
ipython==8.26.0
    # via ipdb
jedi==0.19.1
    # via ipython
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
matplotlib-inline==0.1.7
    # via ipython
packaging==24.1
    # via
    #   pytest
    #   pytest-factoryboy
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pluggy==1.5.0
    # via pytest
prompt-toolkit==3.0.47
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pygments==2.18.0
    # via ipython
pytest==8.3.2
    # via
    #   pytest-cov
    #   pytest-django
    #   pytest-factoryboy
pytest-cov==5.0.0
    # via -r ../requirements.in
pytest-django==4.8.0
    # via -r ../requirements.in
pytest-factoryboy==2.7.0
    # via -r ../requirements.in
python-dateutil==2.9.0.post0
    # via faker
python-dotenv==1.0.1
    # via -r ../requirements.in
pyyaml==6.0.1
    # via drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.19.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   asttokens
    #   python-dateutil
sqlparse==0.5.1
    # via django
stack-data==0.6.3
    # via ipython
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via pytest-factoryboy
uritemplate==4.1.1
    # via drf-spectacular
wcwidth==0.2.13
    # via prompt-toolkit
